In case of an xProfile field requiring options (eg: selectbox, radio buttons), make sure the first option doesn't show a delete link when creating or editing this type of fields.
See #5647

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.0@8739 cdf35c40-ae34-48e0-9cc9-0c9da1808c22

+++ bp-xprofile/bp-xprofile-classes.php
@@ -2839,7 +2839,11 @@ abstract class BP_XProfile_Field_Type {
 							<input type="text" name="<?php echo esc_attr( "{$type}_option[{$j}]" ); ?>" id="<?php echo esc_attr( "{$type}_option{$j}" ); ?>" value="<?php echo esc_attr( stripslashes( $options[$i]->name ) ); ?>" />
 							<input type="<?php echo esc_attr( $control_type ); ?>" name="<?php echo esc_attr( "isDefault_{$type}_option{$default_name}" ); ?>" <?php checked( $options[$i]->is_default_option, true ); ?> value="<?php echo esc_attr( $j ); ?>" />
 							<span><?php _e( 'Default Value', 'buddypress' ); ?></span>
-							<a href="<?php echo esc_url( 'users.php?page=bp-profile-setup&amp;mode=delete_option&amp;option_id=' . $options[$i]->id ); ?>" class="ajax-option-delete" id="delete-<?php echo esc_attr( $options[$i]->id ); ?>">[x]</a>
+
+							<?php if ( 1 <= $i ) : ?>
+								<a href="<?php echo esc_url( 'users.php?page=bp-profile-setup&amp;mode=delete_option&amp;option_id=' . $options[$i]->id ); ?>" class="ajax-option-delete" id="delete-<?php echo esc_attr( $options[$i]->id ); ?>">[x]</a>
+							<?php endif ;?>
+							
 						</p>
 					<?php endfor; ?>
